fix: optimise leave room
