+++ src/views/footerView.jsx
@@ -1,11 +1,8 @@
-import { useState, useCallback, useContext, useRef } from "react";
+import { useState, useCallback, useRef } from "react";
 import {
   useHMSStore,
   ControlBar,
-  ContextMenu,
-  ContextMenuItem,
   AudioPlaylist,
-  HangUpIcon,
   VirtualBackgroundIcon,
   NoiseSupressionIcon,
   Button,
@@ -23,20 +20,19 @@ import {
   selectIsAllowedToPublish,
   selectIsLocalVideoPluginPresent,
   selectIsLocalAudioPluginPresent,
-  selectPermissions,
   selectLocalPeerID,
   selectScreenSharesByPeerId,
   Text,
   selectVideoPlaylist,
   VideoPlaylist,
+  selectIsConnectedToRoom,
 } from "@100mslive/hms-video-react";
-import { useHistory, useParams } from "react-router-dom";
 import { HMSVirtualBackgroundPlugin } from "@100mslive/hms-virtual-background";
 import { HMSNoiseSuppressionPlugin } from "@100mslive/hms-noise-suppression";
-import { AppContext } from "../store/AppContext";
 import { getRandomVirtualBackground } from "../common/utils";
 import { MoreSettings } from "./components/MoreSettings";
 import { AudioVideoToggle } from "./components/AudioVideoToggle";
+import { LeaveRoom } from "./components/LeaveRoom";
 
 export const ConferenceFooter = ({ isChatOpen, toggleChat }) => {
   const isScreenShared = useHMSStore(selectIsLocalScreenShared);
@@ -47,18 +43,13 @@ export const ConferenceFooter = ({ isChatOpen, toggleChat }) => {
     selectIsLocalVideoPluginPresent("@100mslive/hms-virtual-background")
   );
   const hmsActions = useHMSActions();
-  const { isConnected, leave } = useContext(AppContext);
-  const history = useHistory();
-  const params = useParams();
+  const isConnected = useHMSStore(selectIsConnectedToRoom);
+
   const pluginRef = useRef(null);
   const audiopluginRef = useRef(null);
   const isAllowedToPublish = useHMSStore(selectIsAllowedToPublish);
-  const permissions = useHMSStore(selectPermissions);
   const activeVideoPlaylist = useHMSStore(selectVideoPlaylist.selection).id;
-  const [showEndRoomModal, setShowEndRoomModal] = useState(false);
   const [shareAudioModal, setShareAudioModal] = useState(false);
-  const [lockRoom, setLockRoom] = useState(false);
-  const [showMenu, setShowMenu] = useState(false);
 
   const isNoiseSuppression = useHMSStore(
     selectIsLocalAudioPluginPresent("@100mslive/hms-noise-suppression")
@@ -164,15 +155,6 @@ export const ConferenceFooter = ({ isChatOpen, toggleChat }) => {
     [hmsActions]
   );
 
-  function leaveRoom() {
-    leave();
-    if (params.role) {
-      history.push("/leave/" + params.roomId + "/" + params.role);
-    } else {
-      history.push("/leave/" + params.roomId);
-    }
-  }
-
   const leftComponents = [];
   const isAudioScreenshare = !video && !!audio;
 
@@ -228,8 +210,11 @@ export const ConferenceFooter = ({ isChatOpen, toggleChat }) => {
   }
 
   const isPublishing = isAllowedToPublish.video || isAllowedToPublish.audio;
+  if (!isConnected) {
+    return null;
+  }
 
-  return isConnected ? (
+  return (
     <>
       <ControlBar
         leftComponents={leftComponents}
@@ -282,138 +267,13 @@ export const ConferenceFooter = ({ isChatOpen, toggleChat }) => {
           ),
           <MoreSettings key="MoreSettings" />,
         ]}
-        rightComponents={[
-          <ContextMenu
-            classes={{
-              trigger: "w-auto h-auto",
-              root: "static",
-              menu: "w-56 bg-white dark:bg-gray-100",
-              menuItem: "hover:bg-transparent-0 dark:hover:bg-transparent-0",
-            }}
-            onTrigger={value => {
-              if (permissions?.endRoom) {
-                setShowMenu(value);
-              } else {
-                leaveRoom();
-              }
-            }}
-            menuOpen={showMenu}
-            key="LeaveAction"
-            trigger={
-              <Button
-                size="md"
-                shape="rectangle"
-                variant="danger"
-                iconOnly={isMobileDevice()}
-                active={isMobileDevice()}
-                key="LeaveRoom"
-              >
-                <HangUpIcon className={isMobileDevice() ? "" : "mr-2"} />
-                {isMobileDevice() ? "" : "Leave room"}
-              </Button>
-            }
-            menuProps={{
-              anchorOrigin: {
-                vertical: "top",
-                horizontal: "center",
-              },
-              transformOrigin: {
-                vertical: 136,
-                horizontal: "center",
-              },
-            }}
-          >
-            <ContextMenuItem
-              label="Leave Room"
-              key="leaveRoom"
-              classes={{
-                menuTitleContainer: "hidden",
-                menuItemChildren: "my-1 w-full overflow-hidden",
-              }}
-            >
-              <Button
-                shape="rectangle"
-                variant="standard"
-                classes={{ root: "w-full" }}
-                onClick={() => {
-                  leaveRoom();
-                }}
-              >
-                Leave without ending room
-              </Button>
-            </ContextMenuItem>
-
-            {permissions?.endRoom && (
-              <ContextMenuItem
-                label="End Room"
-                key="endRoom"
-                classes={{
-                  menuTitleContainer: "hidden",
-                  menuItemChildren: "my-1 w-full",
-                }}
-              >
-                <Button
-                  shape="rectangle"
-                  variant="danger"
-                  classes={{ root: "w-full" }}
-                  onClick={() => {
-                    setShowEndRoomModal(true);
-                  }}
-                >
-                  End Room for all
-                </Button>
-              </ContextMenuItem>
-            )}
-          </ContextMenu>,
-        ]}
+        rightComponents={[<LeaveRoom key="leaveRoom" />]}
         backgroundButtonOnClick={handleVirtualBackground}
         isBackgroundEnabled={isVBPresent}
       />
       <MessageModal
         {...errorModal}
         onClose={() => setErrorModal(initialModalProps)}
-      />
-      <MessageModal
-        show={showEndRoomModal}
-        onClose={() => {
-          setShowEndRoomModal(false);
-          setLockRoom(false);
-        }}
-        title="End Room"
-        body="Are you sure you want to end the room?"
-        footer={
-          <div className="flex">
-            <div className="flex items-center">
-              <label className="text-base dark:text-white text-gray-100">
-                <input
-                  type="checkbox"
-                  className="mr-1"
-                  onChange={() => setLockRoom(prev => !prev)}
-                  checked={lockRoom}
-                />
-                <span>Lock room</span>
-              </label>
-            </div>
-            <Button
-              classes={{ root: "mr-3 ml-3" }}
-              onClick={() => {
-                setShowEndRoomModal(false);
-                setLockRoom(false);
-              }}
-            >
-              Cancel
-            </Button>
-            <Button
-              variant="danger"
-              onClick={() => {
-                hmsActions.endRoom(lockRoom, "End Room");
-                leaveRoom();
-              }}
-            >
-              End Room
-            </Button>
-          </div>
-        }
       />
       <MessageModal
         show={shareAudioModal}
@@ -451,5 +311,5 @@ export const ConferenceFooter = ({ isChatOpen, toggleChat }) => {
         classes={{ footer: "justify-center", header: "mb-2" }}
       />
     </>
-  ) : null;
+  );
 };
